requirements.txt
Flask==2.0.3
requests==2.26.0
pillow==8.4.0
transformers==4.18.0
flask-cors==3.0.10